+++ frontend/src/features/patients/medical-profile/index.ts
@@ -1,0 +1,1 @@
+export { PatientMedicalProfileForm } from "./MedicalProfileForm";
